chore: engine 패키지명을 play로 변경
- 도메인 특징을 살리도록 이름 변경
