--- src/main/kotlin/yjh/cstar/engine/domain/game/GameFinalizable.kt
@@ -1,6 +1,0 @@
-package yjh.cstar.engine.domain.game
-
-interface GameFinalizable {
-
-    fun finishGame()
-}
