--- src/main/kotlin/yjh/cstar/engine/domain/game/GameInitializable.kt
@@ -1,6 +1,0 @@
-package yjh.cstar.engine.domain.game
-
-interface GameInitializable {
-
-    fun initialize()
-}
