--- src/main/kotlin/yjh/cstar/engine/domain/game/GameRunnable.kt
@@ -1,6 +1,0 @@
-package yjh.cstar.engine.domain.game
-
-interface GameRunnable {
-
-    fun run()
-}
