--- src/main/kotlin/yjh/cstar/engine/domain/quiz/QuizDto.kt
@@ -1,9 +1,0 @@
-package yjh.cstar.engine.domain.quiz
-
-data class QuizDto(val id: Long, val question: String, val answer: String)
-
-fun QuizDto.toModel(): Quiz = Quiz(
-    id = this.id,
-    question = this.question,
-    answer = this.answer
-)
