+++ src/main/kotlin/yjh/cstar/play/domain/game/GameFinalizable.kt
@@ -1,0 +1,6 @@
+package yjh.cstar.play.domain.game
+
+interface GameFinalizable {
+
+    fun finishGame()
+}
