+++ src/main/kotlin/yjh/cstar/play/domain/game/GameInitializable.kt
@@ -1,0 +1,6 @@
+package yjh.cstar.play.domain.game
+
+interface GameInitializable {
+
+    fun initialize()
+}
