+++ src/main/kotlin/yjh/cstar/play/domain/game/GameRunnable.kt
@@ -1,0 +1,6 @@
+package yjh.cstar.play.domain.game
+
+interface GameRunnable {
+
+    fun run()
+}
