+++ src/main/kotlin/yjh/cstar/play/domain/quiz/PlayerAnswer.kt
@@ -1,0 +1,12 @@
+package yjh.cstar.play.domain.quiz
+
+class PlayerAnswer(
+    val answer: String,
+    val quizId: Long,
+    val roomId: Long,
+    val playerId: Long,
+    val nickname: String,
+) {
+
+    fun isCorrect(quiz: Quiz) = quiz.isSameAnswer(answer)
+}
