+++ src/main/kotlin/yjh/cstar/play/domain/ranking/Ranking.kt
@@ -1,0 +1,21 @@
+package yjh.cstar.play.domain.ranking
+
+// <String, Int> "player:1", 10
+class Ranking(
+    private val ranking: LinkedHashMap<String, Int>,
+) {
+
+    companion object {
+        fun of(ranking: List<Pair<String?, Double?>>): Ranking {
+            val mapOfRanking = ranking
+                .filter { it.first != null && it.second != null }
+                .associate { it.first!! to it.second!!.toInt() }
+                .toMap(LinkedHashMap())
+            return Ranking(mapOfRanking)
+        }
+    }
+
+    fun getRanking(): LinkedHashMap<String, Int> = ranking.toList()
+        .sortedByDescending { it.second }
+        .toMap(LinkedHashMap())
+}
