+++ src/test/kotlin/yjh/cstar/redis/RedisConcurrencyTest.kt
@@ -11,12 +11,12 @@ import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
 import org.testcontainers.containers.GenericContainer
 import org.testcontainers.utility.DockerImageName
-import yjh.cstar.engine.domain.quiz.PlayerAnswer
-import yjh.cstar.engine.infrastructure.RedisQueueAnswerProvider
-import yjh.cstar.util.Logger
-import yjh.cstar.util.RedisUtil
 import yjh.cstar.chat.application.PlayerAnswerSendService
 import yjh.cstar.chat.infrastructure.RedisAnswerMessageBroker
+import yjh.cstar.play.domain.quiz.PlayerAnswer
+import yjh.cstar.play.infrastructure.RedisQueueAnswerProvider
+import yjh.cstar.util.Logger
+import yjh.cstar.util.RedisUtil
 import java.util.Collections
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.Executors
